Change indices type on uint16_t in Mesh class

+++ Source/Graphics/Mesh.cpp
@@ -10,7 +10,7 @@ void Mesh::setVertices(const std::vector<float> vertices) {
     this->vertices = vertices;
 }
 
-void Mesh::setIndices(const std::vector<uint32_t> indices) {
+void Mesh::setIndices(const std::vector<uint16_t> indices) {
     this->indices = indices;
 }
